+++ HISTORY.rst
@@ -29,6 +29,7 @@ History
     * Fix briefy.common.utils.data.generate_contextual_slug to be used as default in the BaseMetadata._slug column argument (rudaporto).
     * Change briefy.common.db.mixins.identifiable.GUID.id column to be binary (default and less alocation space) (rudaporto).
     * New attributes in the local roles mixin: relationships and association proxies for easy access to the user permissions. (rudaporto).
+    * New class method Base.__acl__ to map __raw_acl__ to the pyramid format (rudaporto).
  
 
 1.1.2 (2016-10-04)